requirements.txt
python-telegram-bot==20.7
python-telegram-bot[job-queue]
httpx==0.27.0
python-dotenv==1.0.0